+++ src/OrderNotes.php
@@ -60,7 +60,7 @@ class OrderNotes extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '1.1.1';
+    public $schemaVersion = '1.1.2';
 
     // Public Methods
     // =========================================================================